requirements-test.txt
-r requirements.txt
django-debug-toolbar==4.4.2
flake8==7.0.0
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
model-bakery==1.18.0
requests==2.32.3
